+++ native~/Runtime/src/UnityPrepareRendererResources.cpp
@@ -16,11 +16,13 @@
 #include <CesiumShaderProperties.h>
 #include <CesiumUtility/ScopeGuard.h>
 
+#include <DotNet/CesiumForUnity/Cesium3DTileInfo.h>
 #include <DotNet/CesiumForUnity/Cesium3DTileset.h>
 #include <DotNet/CesiumForUnity/CesiumGeoreference.h>
 #include <DotNet/CesiumForUnity/CesiumGlobeAnchor.h>
 #include <DotNet/CesiumForUnity/CesiumMetadata.h>
 #include <DotNet/CesiumForUnity/CesiumObjectPool.h>
+#include <DotNet/CesiumForUnity/CesiumPointCloudRenderer.h>
 #include <DotNet/System/Array1.h>
 #include <DotNet/System/Collections/Generic/List1.h>
 #include <DotNet/System/Object.h>
@@ -324,6 +326,12 @@ void populateMeshDataArray(
           descriptor[numberOfAttributes].dimension = 4;
           descriptor[numberOfAttributes].stream = streamIndex;
           ++numberOfAttributes;
+
+          const int8_t numComponents = gltf.accessors[colorAccessorIt->second]
+                                           .computeNumberOfComponents();
+          if (numComponents == 4) {
+            primitiveInfo.isTranslucent = true;
+          }
         }
 
         // Max number of texture coordinates supported by Unity, see
@@ -542,6 +550,7 @@ struct LoadThreadResult {
   System::Array1<UnityEngine::Mesh> meshes;
   std::vector<CesiumPrimitiveInfo> primitiveInfos{};
 };
+
 } // namespace
 
 UnityPrepareRendererResources::UnityPrepareRendererResources(
@@ -779,6 +788,7 @@ void* UnityPrepareRendererResources::prepareInMainThread(
        showTilesInHierarchy,
        currentOverlayCount,
        &pMetadataComponent,
+       &tile,
        &shaderProperty = _shaderProperty](
           const Model& gltf,
           const Node& node,
@@ -833,6 +843,8 @@ void* UnityPrepareRendererResources::prepareInMainThread(
 
         UnityEngine::MeshFilter meshFilter =
             primitiveGameObject.AddComponent<UnityEngine::MeshFilter>();
+        meshFilter.sharedMesh(unityMesh);
+
         UnityEngine::MeshRenderer meshRenderer =
             primitiveGameObject.AddComponent<UnityEngine::MeshRenderer>();
 
@@ -841,6 +853,7 @@ void* UnityPrepareRendererResources::prepareInMainThread(
 
         UnityEngine::Material opaqueMaterial =
             tilesetComponent.opaqueMaterial();
+
         if (opaqueMaterial == nullptr) {
           if (pMaterial &&
               pMaterial->hasExtension<ExtensionKhrMaterialsUnlit>()) {
@@ -859,10 +872,11 @@ void* UnityPrepareRendererResources::prepareInMainThread(
         material.hideFlags(UnityEngine::HideFlags::HideAndDontSave);
         meshRenderer.material(material);
 
+        bool isTranslucent = primitiveInfo.isTranslucent;
         if (pMaterial) {
           if (pMaterial->pbrMetallicRoughness) {
-            // Add base color factor and metallic-roughness factor regardless of
-            // if the textures are present.
+            // Add base color factor and metallic-roughness factor regardless
+            // of if the textures are present.
             const std::vector<double>& baseColorFactorSrc =
                 pMaterial->pbrMetallicRoughness->baseColorFactor;
             UnityEngine::Vector4 baseColorFactor;
@@ -1015,16 +1029,44 @@ void* UnityPrepareRendererResources::prepareInMainThread(
               0);
         }
 
-        meshFilter.sharedMesh(unityMesh);
+        if (primitiveInfo.containsPoints) {
+          CesiumForUnity::CesiumPointCloudRenderer pointCloudRenderer =
+              primitiveGameObject
+                  .AddComponent<CesiumForUnity::CesiumPointCloudRenderer>();
 
-        if (createPhysicsMeshes &&
-            primitive.mode != MeshPrimitive::Mode::POINTS) {
-          // This should not trigger mesh baking for physics, because the meshes
-          // were already baked in the worker thread.
+          CesiumForUnity::Cesium3DTileInfo tileInfo;
+          tileInfo.usesAdditiveRefinement =
+              tile.getRefine() == Cesium3DTilesSelection::TileRefine::Add;
+          tileInfo.geometricError =
+              static_cast<float>(tile.getGeometricError());
+
+          // TODO: can we make AccessorView retrieve the min/max for us?
+          const Accessor* pPositionAccessor =
+              Model::getSafe(&gltf.accessors, positionAccessorID);
+          glm::vec3 min(
+              pPositionAccessor->min[0],
+              pPositionAccessor->min[1],
+              pPositionAccessor->min[2]);
+          glm::vec3 max(
+              pPositionAccessor->max[0],
+              pPositionAccessor->max[1],
+              pPositionAccessor->max[2]);
+          glm::vec3 dimensions(transform * glm::dvec4(max - min, 0));
+
+          tileInfo.dimensions =
+              UnityEngine::Vector3{dimensions.x, dimensions.y, dimensions.z};
+          tileInfo.isTranslucent = primitiveInfo.isTranslucent;
+          pointCloudRenderer.tileInfo(tileInfo);
+        }
+
+        if (createPhysicsMeshes && !primitiveInfo.containsPoints) {
+          // This should not trigger mesh baking for physics, because the
+          // meshes were already baked in the worker thread.
           UnityEngine::MeshCollider meshCollider =
               primitiveGameObject.AddComponent<UnityEngine::MeshCollider>();
           meshCollider.sharedMesh(unityMesh);
         }
+
         const ExtensionMeshPrimitiveExtFeatureMetadata* pMetadata =
             primitive.getExtension<ExtensionMeshPrimitiveExtFeatureMetadata>();
         if (pMetadata) {